fixed several MSVC warnings + 2 fails in tests
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@2546 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ tests/cxts/cxts.cpp
@@ -958,7 +958,7 @@ CvBadArgTest::~CvBadArgTest()
 
 int CvBadArgTest::run_test_case( int expected_code, const char* descr )
 {
-    double new_t = cv::getTickCount(), dt;
+    double new_t = (double)cv::getTickCount(), dt;
     if( test_case_idx < 0 )
     {
         test_case_idx = 0;
